feat: cadastro via Telegram e comando /receber
- Migrations 008-010: cpf/phone em users, telegram_conversations, revenues
- Fluxo guiado de cadastro via /start: nome → cpf → email → telefone → cria conta
- Qualquer mensagem de usuário não cadastrado redireciona para /start
- /receber <valor> <descrição> registra receita vinculada ao usuário
- RevenueUseCase + UserUseCase com RegisterFromTelegram e state machine de conversa
- Fix: LastMigration int64→int32 alinhado com schema gerado pelo sqlc

+++ internal/infrastructure/pgstore/postgres.sql.go
@@ -20,7 +20,7 @@ FROM (SELECT 1) AS dummy
 type GetDBStatusRow struct {
 	Version        string `json:"version"`
 	MaxConnections int32  `json:"max_connections"`
-	LastMigration  int64  `json:"last_migration"`
+	LastMigration  int32  `json:"last_migration"`
 }
 
 func (q *Queries) GetDBStatus(ctx context.Context) (GetDBStatusRow, error) {
